add unit tests

+++ Database/src/products/products.controller.ts
@@ -12,25 +12,35 @@ import {
 import { ProductsService } from './products.service';
 import { CreateProductDto } from './dto/create-product.dto';
 import { UpdateProductDto } from './dto/update-product.dto';
-import { Product as ProductModel } from '@prisma/client';
+import { Product } from '@prisma/client';
 
 @Controller('products')
 export class ProductsController {
   constructor(private readonly productsService: ProductsService) {}
 
   @Get()
-  async findAll(): Promise<ProductModel[]> {
+  async findAll(): Promise<Product[]> {
     return await this.productsService.findAll();
   }
 
   @Get(':id')
-  async findOne(@Param('id') id: string): Promise<ProductModel> {
+  async findOne(@Param('id') id: string): Promise<Product> {
     const idNumber = Number(id);
-    return await this.productsService.findOne(idNumber);
+    if (isNaN(idNumber)) {
+      throw new BadRequestException(
+        'Invalid ID format. Please provide a valid number.',
+      );
+    }
+
+    const existedProduct = await this.productsService.findOne(idNumber);
+    if (existedProduct) {
+      return existedProduct;
+    }
+    throw new NotFoundException('Product not found');
   }
 
   @Post()
-  async create(@Body() createDto: CreateProductDto): Promise<ProductModel> {
+  async create(@Body() createDto: CreateProductDto): Promise<Product> {
     const existedProduct = await this.productsService.findByName(
       createDto.name,
     );
@@ -44,24 +54,37 @@ export class ProductsController {
   async update(
     @Param('id') id: string,
     @Body() updateDto: UpdateProductDto,
-  ): Promise<ProductModel> {
+  ): Promise<Product> {
     const idNumber = Number(id);
+    if (isNaN(idNumber)) {
+      throw new BadRequestException(
+        'Invalid ID format. Please provide a valid number.',
+      );
+    }
+
     const existedProduct = await this.productsService.findOne(idNumber);
     if (existedProduct) {
+      existedProduct.updatedAt = new Date();
       return await this.productsService.update(idNumber, updateDto);
     } else {
-      throw new NotFoundException();
+      throw new NotFoundException('Product not found');
     }
   }
 
   @Delete(':id')
-  async remove(@Param('id') id: string): Promise<ProductModel> {
+  async remove(@Param('id') id: string): Promise<Product> {
     const idNumber = Number(id);
+    if (isNaN(idNumber)) {
+      throw new BadRequestException(
+        'Invalid ID format. Please provide a valid number.',
+      );
+    }
+
     const existedProduct = await this.productsService.findOne(idNumber);
     if (existedProduct) {
       return this.productsService.remove(idNumber);
     } else {
-      throw new NotFoundException();
+      throw new NotFoundException('Product not found');
     }
   }
 }
